edit

+++ admission_app/views.py
@@ -69,7 +69,8 @@ def admin(request):
 
     user = User.objects.get(id=request.session["userId"])
     courses =Course.objects.all()
-    students = User.objects.filter(state="Pennding")
+    students = User.objects.filter(state="pennding")
+    request.session["request_pennding"]=len(students)
     if(user.role=="admin"):
         context = {
             "user": user,
@@ -80,6 +81,8 @@ def admin(request):
     return HttpResponse("Please authenticate first")
 
 def add_course(request):
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     if request.method == "POST":
         name = request.POST["name"]
         desc = request.POST["desc"]
@@ -90,23 +93,28 @@ def add_course(request):
     return render(request,"add_course.html")
 
 def edit_state(request,id,state):
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     user = User.objects.get(id=id)
-    if user.course.capacity == len(user.course.users.all().filter(state='approve')):
+    if state=='decline':
+        user.course=None
+    elif user.course.capacity == len(user.course.users.all().filter(state='approve')):
         messages.error(request,f'{user.course.name} course is full')
         #state var here is approve, we need to change it pennding
         state=user.state
     user.state = state
     user.save()
-    if state=='decline':
-        user.course=None
     return redirect('/admin')
 
 def apply_course(request,id):
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     if request.method =='POST':
         user = User.objects.get(id=request.session["userId"])
         if (not user.course):
             course=Course.objects.get(id=id)
             user.course=course
+            user.state="pennding"
             user.save()
             return redirect(f'/student_profile/{user.id}')
         else: #user already applied to another course 
@@ -114,7 +122,8 @@ def apply_course(request,id):
     return redirect(f'/student_profile/{user.id}')
 
 def show_student(request,id):
-    
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     if request.method=='POST':
         user=User.objects.get(id=id)
         user.first_name=request.POST['first_name']
@@ -131,15 +140,24 @@ def show_student(request,id):
     return render(request,'student_profile.html',context)
 
 def delete_course(request,id):
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     Course.objects.get(id=id).delete()
     return redirect('/admin')
 
 def edit_course(request,id):
+    if "userId" not in request.session:
+        return HttpResponse("Please authenticate first")
     course = Course.objects.get(id=id)
     if request.method == "POST":
         course.name = request.POST["name"]
         course.desc = request.POST["desc"]
-        course.capacity=request.POST["capacity"]
+        if len(course.users.filter(state="approve"))<int(request.POST["capacity"]):
+          course.capacity=request.POST["capacity"]
+        else:
+            messages.error(request,"The new seats number is less than the number of applicants.")
+            return redirect(f'/edit_course/{course.id}')
+
         if request.FILES.get('photo'):
             course.photo = request.FILES['photo']
         course.save()
@@ -154,18 +172,17 @@ def edit_course(request,id):
 def edit_profile(request):
     this_user=User.objects.get(id=request.session["userId"])
     if request.method == "POST":
-        course_id=request.POST["course"]
-        this_course=Course.objects.get(id=course_id)
         this_user.first_name=request.POST["first_name"]
         this_user.last_name=request.POST["last_name"]
-        this_user.course=this_course
+        if "course" in request.POST:
+          course_id=request.POST["course"]
+          this_course=Course.objects.get(id=course_id)
+          this_user.course=this_course
         this_user.save()
         return redirect(f'/student_profile/{this_user.id}')
     
     return redirect(f'/student_profile/{this_user.id}')
 
-
-    
 def logout(request):
     request.session.clear()
     # messages.success(request, "You have been logged out!")
